Addressed PR comments.
