+++ psl-core/src/main/java/org/linqs/psl/database/rdbms/RDBMSDatabase.java
@@ -84,88 +84,21 @@ public class RDBMSDatabase extends Database {
 	private static final double DEFAULT_UNOBSERVED_VALUE = 0.0;
 
 	/**
-	 * The backing data store that created this database.
-	 * Connection are obtained from here.
-	 */
-	private final RDBMSDataStore parentDataStore;
-
-	/**
-	 * The partition ID in which this database writes.
-	 */
-	private final Partition writePartition;
-	private final int writeID;
-
-	/**
-	 * The partition IDs that this database reads from.
-	 */
-	private final List<Partition> readPartitions;
-	private final List<Integer> readIDs;
-
-	/**
 	 * Predicates that, for the purpose of this database, are closed.
 	 */
 	private final Set<Predicate> closedPredicates;
 
-	/**
-	 * Mapping from a predicate to its database handle.
-	 */
-	private final Map<Predicate, PredicateInfo> predicates;
-
-	/**
-	 * The atom cache for this database.
-	 */
-	private final AtomCache cache;
-
-	/**
-	 * Keeps track of the open / closed status of this database.
-	 */
-	private boolean closed;
-
 	public RDBMSDatabase(RDBMSDataStore parent,
 			Partition write, Partition[] read,
-			Map<Predicate, PredicateInfo> predicates,
 			Set<StandardPredicate> closed) {
-		this.parentDataStore = parent;
-		this.writePartition = write;
-		this.writeID = write.getID();
-
-		this.readPartitions = Arrays.asList(read);
-		this.readIDs = new ArrayList<Integer>(read.length);
-		for (int i = 0; i < read.length; i ++) {
-			this.readIDs.add(read[i].getID());
-		}
-
-		if (!this.readIDs.contains(writeID)) {
-			this.readIDs.add(writeID);
-		}
-
-		this.predicates = new HashMap<Predicate, PredicateInfo>();
-		for (Map.Entry<Predicate, PredicateInfo> entry : predicates.entrySet()) {
-			this.predicates.put(entry.getKey(), entry.getValue());
-		}
+		super(parent, write, read);
 
 		this.closedPredicates = new HashSet<Predicate>();
 		if (closed != null) {
 			this.closedPredicates.addAll(closed);
 		}
 
-		this.cache = new AtomCache(this);
-
 		this.closed = false;
-	}
-
-	/**
-	 * Helper method for getting a predicate handle
-	 * @param predicate	The predicate to lookup
-	 * @return	The handle associated with the predicate
-	 */
-	public PredicateInfo getPredicateInfo(Predicate predicate) {
-		PredicateInfo info = predicates.get(predicate);
-		if (info == null) {
-			throw new IllegalArgumentException("Predicate not registered with database.");
-		}
-
-		return info;
 	}
 
 	@Override
@@ -212,7 +145,7 @@ public class RDBMSDatabase extends Database {
 
 		try (
 			Connection connection = getConnection();
-			PreparedStatement statement = getAtomDelete(connection, getPredicateInfo(atom.getPredicate()), atom.getArguments());
+			PreparedStatement statement = getAtomDelete(connection, ((RDBMSDataStore)parentDataStore).getPredicateInfo(atom.getPredicate()), atom.getArguments());
 		) {
 			if (statement.executeUpdate() > 0) {
 				return true;
@@ -222,74 +155,6 @@ public class RDBMSDatabase extends Database {
 		} catch (SQLException ex) {
 			throw new RuntimeException("Error deleting atom: " + atom, ex);
 		}
-	}
-
-	@Override
-	public int countAllGroundAtoms(StandardPredicate predicate) {
-		List<Integer> partitions = new ArrayList<Integer>();
-		partitions.addAll(readIDs);
-		partitions.add(writeID);
-
-		return countAllGroundAtoms(predicate, partitions);
-	}
-
-	@Override
-	public int countAllGroundRandomVariableAtoms(StandardPredicate predicate) {
-		// Closed predicates have no random variable atoms.
-		if (isClosed(predicate)) {
-			return 0;
-		}
-
-		// All the atoms should be random vairable, since we are pulling from the write parition of an open predicate.
-		List<Integer> partitions = new ArrayList<Integer>(1);
-		partitions.add(writeID);
-
-		return countAllGroundAtoms(predicate, partitions);
-	}
-
-	@Override
-	public List<GroundAtom> getAllGroundAtoms(StandardPredicate predicate) {
-		List<Integer> partitions = new ArrayList<Integer>();
-		partitions.addAll(readIDs);
-		partitions.add(writeID);
-
-		return getAllGroundAtoms(predicate, partitions);
-	}
-
-	@Override
-	public List<RandomVariableAtom> getAllGroundRandomVariableAtoms(StandardPredicate predicate) {
-		// Closed predicates have no random variable atoms.
-		if (isClosed(predicate)) {
-			return new ArrayList<RandomVariableAtom>();
-		}
-
-		// All the atoms should be random vairable, since we are pulling from the write parition of an open predicate.
-		List<Integer> partitions = new ArrayList<Integer>(1);
-		partitions.add(writeID);
-		List<GroundAtom> groundAtoms = getAllGroundAtoms(predicate, partitions);
-
-		List<RandomVariableAtom> atoms = new ArrayList<RandomVariableAtom>(groundAtoms.size());
-		for (GroundAtom atom : groundAtoms) {
-			atoms.add((RandomVariableAtom)atom);
-		}
-
-		return atoms;
-	}
-
-	@Override
-	public List<ObservedAtom> getAllGroundObservedAtoms(StandardPredicate predicate) {
-		// Note that even open predicates may have observed atoms (partially observed predicates).
-
-		// Only pull from the read partitions.
-		List<GroundAtom> groundAtoms = getAllGroundAtoms(predicate, readIDs);
-
-		// All the atoms will be observed since we are pulling from only read partitions.
-		List<ObservedAtom> atoms = new ArrayList<ObservedAtom>(groundAtoms.size());
-		for (GroundAtom atom : groundAtoms) {
-			atoms.add((ObservedAtom)atom);
-		}
-
-		return atoms;
 	}
 
 	@Override
@@ -317,7 +182,7 @@ public class RDBMSDatabase extends Database {
 		try (Connection connection = getConnection()) {
 			// Upsert each predicate batch.
 			for (Map.Entry<Predicate, List<RandomVariableAtom>> entry : atomsByPredicate.entrySet()) {
-				try (PreparedStatement statement = getAtomUpsert(connection, getPredicateInfo(entry.getKey()))) {
+				try (PreparedStatement statement = getAtomUpsert(connection, ((RDBMSDataStore)parentDataStore).getPredicateInfo(entry.getKey()))) {
 					int batchSize = 0;
 
 					// Set all the upsert params.
@@ -359,15 +224,8 @@ public class RDBMSDatabase extends Database {
 	}
 
 	@Override
-	public void commit(RandomVariableAtom atom) {
-		List<RandomVariableAtom> atoms = new ArrayList<RandomVariableAtom>(1);
-		atoms.add(atom);
-		commit(atoms);
-	}
-
-	@Override
 	public void moveToWritePartition(StandardPredicate predicate, int oldPartitionId) {
-		PredicateInfo predicateInfo = getPredicateInfo(predicate);
+		PredicateInfo predicateInfo = ((RDBMSDataStore)parentDataStore).getPredicateInfo(predicate);
 
 		try (
 			Connection connection = getConnection();
@@ -462,21 +320,8 @@ public class RDBMSDatabase extends Database {
 		return closedPredicates.contains(predicate);
 	}
 
-	@Override
-	public DataStore getDataStore() {
-		return parentDataStore;
-	}
-
-	public List<Partition> getReadPartitions() {
-		return Collections.unmodifiableList(readPartitions);
-	}
-
-	public Partition getWritePartition() {
-		return writePartition;
-	}
-
 	private Connection getConnection() {
-		return parentDataStore.getConnection();
+		return ((RDBMSDataStore)parentDataStore).getConnection();
 	}
 
 	@Override
@@ -485,7 +330,7 @@ public class RDBMSDatabase extends Database {
 			throw new IllegalStateException("Cannot close database after it has been closed.");
 		}
 
-		parentDataStore.releasePartitions(this);
+		((RDBMSDataStore)parentDataStore).releasePartitions(this);
 		closed = true;
 	}
 
@@ -504,7 +349,7 @@ public class RDBMSDatabase extends Database {
 	}
 
 	private PreparedStatement getAtomUpsert(Connection connection, PredicateInfo predicate) {
-		return predicate.createUpsertStatement(connection, parentDataStore.getDriver());
+		return predicate.createUpsertStatement(connection, ((RDBMSDataStore)parentDataStore).getDriver());
 	}
 
 	private PreparedStatement getAtomDelete(Connection connection, PredicateInfo predicate, Term[] arguments) {
@@ -586,7 +431,8 @@ public class RDBMSDatabase extends Database {
 	/**
 	 * @param create Create an atom if one does not exist.
 	 */
-	private GroundAtom getAtom(StandardPredicate predicate, boolean create, Constant... arguments) {
+	@Override
+	public GroundAtom getAtom(StandardPredicate predicate, boolean create, Constant... arguments) {
 		QueryAtom queryAtom = new QueryAtom(predicate, arguments);
 		GroundAtom result = cache.getCachedAtom(queryAtom);
 		if (result != null) {
@@ -601,7 +447,7 @@ public class RDBMSDatabase extends Database {
 	 */
 	private GroundAtom fetchAtom(StandardPredicate predicate, boolean create, Constant... arguments) {
 		// Ensure this database has this predicate.
-		getPredicateInfo(predicate);
+		((RDBMSDataStore)parentDataStore).getPredicateInfo(predicate);
 
 		GroundAtom result = queryDBForAtom(predicate, arguments);
 
@@ -625,8 +471,8 @@ public class RDBMSDatabase extends Database {
 	 */
 	private GroundAtom queryDBForAtom(StandardPredicate predicate, Constant[] arguments) {
 		try (
-			Connection conn = parentDataStore.getConnection();
-			PreparedStatement statement = getAtomQuery(conn, predicates.get(predicate), arguments);
+			Connection conn = getConnection();
+			PreparedStatement statement = getAtomQuery(conn, ((RDBMSDataStore)parentDataStore).getPredicateInfo(predicate), arguments);
 			ResultSet resultSet = statement.executeQuery();
 		) {
 			if (!resultSet.next()) {
@@ -645,13 +491,10 @@ public class RDBMSDatabase extends Database {
 		}
 	}
 
-	public boolean hasAtom(StandardPredicate predicate, Constant... arguments) {
-		return getAtom(predicate, false, arguments) != null;
-	}
-
-	private List<GroundAtom> getAllGroundAtoms(StandardPredicate predicate, List<Integer> partitions) {
+	@Override
+	public List<GroundAtom> getAllGroundAtoms(StandardPredicate predicate, List<Integer> partitions) {
 		List<GroundAtom> atoms = new ArrayList<GroundAtom>();
-		PredicateInfo predicateInfo = getPredicateInfo(predicate);
+		PredicateInfo predicateInfo = ((RDBMSDataStore)parentDataStore).getPredicateInfo(predicate);
 
 		// Columns for each argument to the predicate.
 		List<String> argumentCols = predicateInfo.argumentColumns();
@@ -677,8 +520,9 @@ public class RDBMSDatabase extends Database {
 		return atoms;
 	}
 
-	private int countAllGroundAtoms(StandardPredicate predicate, List<Integer> partitions) {
-		PredicateInfo predicateInfo = getPredicateInfo(predicate);
+	@Override
+	public int countAllGroundAtoms(StandardPredicate predicate, List<Integer> partitions) {
+		PredicateInfo predicateInfo = ((RDBMSDataStore)parentDataStore).getPredicateInfo(predicate);
 
 		try (
 			Connection connection = getConnection();
